query: consolidate multiple files
